pothos-library: fixed message back-pressure mechanism
The same buffer manager is used to provide "tokens" for message back-pressure,
however, due to a optimization to not pop the buffer under partial consumption,
the tokens/buffers were never being removed from the manager.

Fixed it with some additional logic (len != 0) and an assertion that
the popped token is unique (not still inside the manager).