pothos-library: switched serialization to portable text archive
We have successfully tested the remote stuff mixing architectures arm + x64.
The serialization for object is now text archive because binary wasnt portable.
Also there was a fix for the remote proxy to store the lookup TID value as a fixed size integer,
otherwise the tid could get truncated if it was a size_t which varies across architecture.