pothos-blocks: Network blocks support jumbo sized transfers
This addresses a bug where really large transfers fail.
The problem was that the send size was too large for the send call.
The simple fix was to call send() in a loop to complete the transfer.

In addition, the header format was changed to accommodate larger transfers.
The 16 bit byte count was easily too small for large transfers,
like the payload for the waveform plotter when its zoomed way out.
Both the packet count and the byte count are now 32-bit words,
and the magic header word has been version bumped to PTH2
so that communication between versions fails "nicely".