Fixed weak storage of plugin's associated module
Plugin was storing a pointer to the associated module.
The problem is that the pointer was to a temporary object,
and not to the underlying storage container for the module.
Replace the pointer with a weak ptr to the internal container.

This fix caused an ABI change and has been put on master.
Its mostly an aesthetic fix for the PothosUtil plugin prints
and the PothosGui host explorer for plugins.