Callable and ManagedClass now 100%  variadic
Replaced the generated stuff with overloads for 0 arguments
and overloads for leading variadic argument pack.

First work around:
The overloads with zero arguments are needed to specify
overloads that have no arguments using the return type.

There is a minor annoying when switching to variadics here,
its not possible to specify some of the arguments
or all of the arguments and the return type,
or the arguments when the call is templated.

Second work around:
Specify nothing, or specify the entire argument pack
without the return or class type. Templated calls
need to have their template specified on the call itself.

Doing away with generated headers and working with
this inconvenience of specifying extra arguments
sometimes is hopefully worth the effort.