Simplify buffer chunk by removing extra shared buffer reference
This makes buffer chunk lighter-weight to copy and move around.
The managed buffer holds a shared buffer reference inside of it
and we can simply reuse this reference. The main change is to make
a managed buffer with no manager when only a shared buffer was required.
Some changed to managed buffer followed suit to make the calls
work it was also null (such as getBuffer, useCount, unique).
This simplifies the implementation of BufferChunk as well.