BufferAccumulator require() fixes for circbuf and for empty condition.
When the accumulator was empty, a pool chunk would be added, then a
zero-length memcpy performed. Instead, return if empty.

With circbuf with numBuffers > 1, backpressure is implemented by filling
the accumulator queue with empty chunks whose contents have been
accumulated into the front chunk. In this case, the function should
return.